expression evaluator: fixed to strictly 32 bit value internally, as docs claimed
This bugfix may change values produced by shift operations compared to
older versions of sjasmplus (not with common ordinary Z80 values near
16b range, only when values used were near the 32/64 bit limits).

In such case also older versions of sjasmplus could have produced
different machine code depending on which binary of sjasmplus was used
(32b vs 64b and windows vs linux).

After this change all sjasmplus versions should produce strictly 32b
signed expression values, in accordance with original documentation
(which I did make even more explicit than before).

+++ sjasm/parser.h
@@ -44,7 +44,7 @@ void SetLastParsedLabel(const char* label);
 void ParseLine(bool = true);
 void ParseLineSafe(bool = true);
 void ParseStructLine(CStructure* st);
-unsigned long LuaCalculate(char *str);
+uint32_t LuaCalculate(char *str);
 void LuaParseLine(char *str);
 void LuaParseCode(char *str);
 //eof parser.h
